+++ src/main/java/com/besson/endfield/block/custom/ElectricPylonBlock.java
@@ -1,10 +1,14 @@
 package com.besson.endfield.block.custom;
 
 import com.besson.endfield.block.ModBlockEntityWithFacing;
+import com.besson.endfield.blockentity.ModBlockEntities;
 import com.besson.endfield.blockentity.custom.ElectricPylonBlockEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class ElectricPylonBlock extends ModBlockEntityWithFacing {
@@ -15,5 +19,12 @@ public class ElectricPylonBlock extends ModBlockEntityWithFacing {
     @Override
     public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
         return new ElectricPylonBlockEntity(pos, state);
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return checkType(type, ModBlockEntities.ELECTRIC_PYLON,
+                (world1, pos, state1, blockEntity) ->
+                    ElectricPylonBlockEntity.tick(world1, pos, state1, (ElectricPylonBlockEntity) blockEntity));
     }
 }