+++ src/main/java/com/besson/endfield/block/custom/RelayTowerBlock.java
@@ -1,11 +1,15 @@
 package com.besson.endfield.block.custom;
 
 import com.besson.endfield.block.ModBlockEntityWithFacing;
+import com.besson.endfield.blockentity.ModBlockEntities;
 import com.besson.endfield.blockentity.custom.RelayTowerBlockEntity;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public class RelayTowerBlock extends ModBlockEntityWithFacing {
@@ -17,5 +21,12 @@ public class RelayTowerBlock extends ModBlockEntityWithFacing {
     @Override
     public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
         return new RelayTowerBlockEntity(pos, state);
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return checkType(type, ModBlockEntities.RELAY_TOWER,
+                (world1, pos, state1, blockEntity) ->
+                        RelayTowerBlockEntity.tick(world1, pos, state1, (RelayTowerBlockEntity) blockEntity));
     }
 }