+++ src/main/java/com/besson/endfield/blockentity/custom/RefiningUnitBlockEntity.java
@@ -1,19 +1,80 @@
 package com.besson.endfield.blockentity.custom;
 
+import com.besson.endfield.block.ElectrifiableDevice;
+import com.besson.endfield.blockentity.ImplementedInventory;
 import com.besson.endfield.blockentity.ModBlockEntities;
+import com.besson.endfield.recipe.custom.RefiningUnitRecipe;
+import com.besson.endfield.screen.custom.RefiningUnitScreenHandler;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.Inventories;
+import net.minecraft.inventory.SimpleInventory;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.listener.ClientPlayPacketListener;
+import net.minecraft.network.packet.Packet;
+import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
+import net.minecraft.screen.PropertyDelegate;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.Text;
+import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
 import software.bernie.geckolib.animatable.GeoBlockEntity;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.core.animation.AnimatableManager;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
-public class RefiningUnitBlockEntity extends BlockEntity implements GeoBlockEntity {
+import java.util.Optional;
+
+public class RefiningUnitBlockEntity extends BlockEntity implements GeoBlockEntity, ExtendedScreenHandlerFactory, ImplementedInventory, ElectrifiableDevice {
+    private final DefaultedList<ItemStack> inventory = DefaultedList.ofSize(2, ItemStack.EMPTY);
+    public static final int INPUT_SLOT = 0;
+    public static final int OUTPUT_SLOT = 1;
+
+    protected final PropertyDelegate propertyDelegate;
+    private int progress = 0;
+    private int maxProgress = 80;
+
+    private int storePower = 0;
+    private static final int POWER_PRE_TICK = 5;
+    private boolean isWorking = false;
+
     private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
 
     public RefiningUnitBlockEntity(BlockPos pos, BlockState state) {
         super(ModBlockEntities.REFINING_UNIT, pos, state);
+        this.propertyDelegate = new PropertyDelegate() {
+            @Override
+            public int get(int index) {
+                return switch (index) {
+                    case 0 -> RefiningUnitBlockEntity.this.progress;
+                    case 1 -> RefiningUnitBlockEntity.this.maxProgress;
+                    default -> 0;
+                };
+            }
+
+            @Override
+            public void set(int index, int value) {
+                switch (index) {
+                    case 0 -> RefiningUnitBlockEntity.this.progress = value;
+                    case 1 -> RefiningUnitBlockEntity.this.maxProgress = value;
+                }
+            }
+
+            @Override
+            public int size() {
+                return 2;
+            }
+        };
     }
 
     @Override
@@ -24,5 +85,152 @@ public class RefiningUnitBlockEntity extends BlockEntity implements GeoBlockEnti
     @Override
     public AnimatableInstanceCache getAnimatableInstanceCache() {
         return cache;
+    }
+
+    @Override
+    public DefaultedList<ItemStack> getItems() {
+        return inventory;
+    }
+
+    @Override
+    public void writeScreenOpeningData(ServerPlayerEntity serverPlayerEntity, PacketByteBuf packetByteBuf) {
+        packetByteBuf.writeBlockPos(this.pos);
+    }
+
+    @Override
+    public Text getDisplayName() {
+        return Text.translatable("blockEntity.refining_unit");
+    }
+
+    @Override
+    public @Nullable ScreenHandler createMenu(int syncId, PlayerInventory playerInventory, PlayerEntity player) {
+        return new RefiningUnitScreenHandler(syncId, playerInventory, this, this.propertyDelegate);
+    }
+
+    @Override
+    public @Nullable Packet<ClientPlayPacketListener> toUpdatePacket() {
+        return BlockEntityUpdateS2CPacket.create(this);
+    }
+
+    @Override
+    public NbtCompound toInitialChunkDataNbt() {
+        return this.createNbt();
+    }
+
+    @Override
+    public void readNbt(NbtCompound nbt) {
+        super.readNbt(nbt);
+        Inventories.readNbt(nbt, inventory);
+        this.progress = nbt.getInt("progress");
+    }
+
+    @Override
+    protected void writeNbt(NbtCompound nbt) {
+        super.writeNbt(nbt);
+        Inventories.writeNbt(nbt, inventory);
+        nbt.putInt("progress", progress);
+    }
+
+    public static void tick(World world, BlockPos pos, BlockState state, RefiningUnitBlockEntity be) {
+        if (world.isClient()) return;
+
+        if (be.needsPower() && be.isWorking) {
+            be.isWorking = false;
+            be.markDirty();
+        } else if (!be.needsPower() && !be.isWorking) {
+            be.isWorking = true;
+            be.markDirty();
+        }
+
+        if (be.isOutputSlotAvailable() && be.isWorking) {
+            if (be.hasCorrectRecipe(world) && be.storePower >= POWER_PRE_TICK) {
+                be.storePower -= POWER_PRE_TICK;
+                be.increaseProgress();
+                be.markDirty();
+
+                if (be.hasCraftingFinished()) {
+                    be.craftItem(world);
+                    be.resetProgress();
+                    be.markDirty();
+                }
+            } else {
+                be.resetProgress();
+                be.markDirty();
+            }
+        } else {
+            be.resetProgress();
+            be.markDirty();
+        }
+    }
+
+    private void resetProgress() {
+        this.progress = 0;
+    }
+
+    private void craftItem(World world) {
+
+        Optional<RefiningUnitRecipe> match = getMatchRecipe(world);
+
+        if (match.isPresent()) {
+            ItemStack result = match.get().getOutput(world.getRegistryManager());
+            this.setStack(OUTPUT_SLOT,
+                    new ItemStack(result.getItem(), getStack(OUTPUT_SLOT).getCount() + result.getCount()));
+            this.removeStack(INPUT_SLOT, 1);
+        }
+
+    }
+
+    private Optional<RefiningUnitRecipe> getMatchRecipe(World world) {
+        SimpleInventory inv = new SimpleInventory(this.size());
+        for (int i = 0; i < this.size(); i++) {
+            inv.setStack(i, this.getStack(i));
+        }
+        return world.getRecipeManager()
+                .getFirstMatch(RefiningUnitRecipe.Type.INSTANCE, inv, world);
+    }
+
+    private boolean hasCraftingFinished() {
+        return progress >= maxProgress;
+    }
+
+    private void increaseProgress() {
+        this.progress++;
+    }
+
+    private boolean hasCorrectRecipe(World world) {
+        Optional<RefiningUnitRecipe> match = getMatchRecipe(world);
+
+        if (match.isPresent()) {
+            ItemStack result = match.get().getOutput(world.getRegistryManager());
+            return canInsertAmountIntoOutputSlot(result) && canInsertItemIntoOutputSlot(result.getItem());
+        }
+
+        return false;
+    }
+    private boolean canInsertAmountIntoOutputSlot(ItemStack result) {
+        return this.getStack(OUTPUT_SLOT).getCount() + result.getCount() <= getStack(OUTPUT_SLOT).getMaxCount();
+    }
+
+    private boolean canInsertItemIntoOutputSlot(Item item) {
+        return this.getStack(OUTPUT_SLOT).getItem() == item || this.getStack(OUTPUT_SLOT).isEmpty();
+    }
+
+    private boolean isOutputSlotAvailable() {
+        return this.getStack(OUTPUT_SLOT).isEmpty() || this.getStack(OUTPUT_SLOT).getCount() < this.getStack(OUTPUT_SLOT).getMaxCount();
+    }
+
+    @Override
+    public void receiveElectricCharge(int amount) {
+        this.storePower += amount;
+    }
+
+    @Override
+    public boolean needsPower() {
+        return this.storePower < POWER_PRE_TICK;
+    }
+
+    @Override
+    public int getRequiredPower() {
+        return POWER_PRE_TICK;
     }
 }