+++ src/main/java/com/besson/endfield/recipe/ModRecipes.java
@@ -2,6 +2,7 @@ package com.besson.endfield.recipe;
 
 import com.besson.endfield.ArknightsEndfield;
 import com.besson.endfield.recipe.custom.PortableOriginiumRigRecipe;
+import com.besson.endfield.recipe.custom.RefiningUnitRecipe;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
@@ -12,5 +13,10 @@ public class ModRecipes {
                 PortableOriginiumRigRecipe.Serializer.INSTANCE);
         Registry.register(Registries.RECIPE_TYPE, new Identifier(ArknightsEndfield.MOD_ID, PortableOriginiumRigRecipe.Type.ID),
                 PortableOriginiumRigRecipe.Type.INSTANCE);
+
+        Registry.register(Registries.RECIPE_SERIALIZER, new Identifier(ArknightsEndfield.MOD_ID, RefiningUnitRecipe.Serializer.ID),
+                RefiningUnitRecipe.Serializer.INSTANCE);
+        Registry.register(Registries.RECIPE_TYPE, new Identifier(ArknightsEndfield.MOD_ID, RefiningUnitRecipe.Type.ID),
+                RefiningUnitRecipe.Type.INSTANCE);
     }
 }