+++ src/main/java/com/besson/endfield/screen/custom/ThermalBankScreenHandler.java
@@ -1,0 +1,94 @@
+package com.besson.endfield.screen.custom;
+
+import com.besson.endfield.blockentity.custom.ThermalBankBlockEntity;
+import com.besson.endfield.screen.ModScreens;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.Inventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.screen.ArrayPropertyDelegate;
+import net.minecraft.screen.PropertyDelegate;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.screen.slot.Slot;
+
+public class ThermalBankScreenHandler extends ScreenHandler {
+    private final Inventory inv;
+    private final PropertyDelegate propertyDelegate;
+    public final ThermalBankBlockEntity entity;
+
+    public  ThermalBankScreenHandler(int syncId, PlayerInventory playerInventory, PacketByteBuf buf) {
+        this(syncId, playerInventory, playerInventory.player.getWorld().getBlockEntity(buf.readBlockPos()), new ArrayPropertyDelegate(2));
+    }
+
+    public ThermalBankScreenHandler(int syncId, PlayerInventory playerInventory, BlockEntity blockEntity, PropertyDelegate propertyDelegate) {
+        super(ModScreens.THERMAL_BANK_SCREEN, syncId);
+        checkSize((Inventory) blockEntity, 1);
+        this.inv = (Inventory) blockEntity;
+        inv.onOpen(playerInventory.player);
+        this.propertyDelegate = propertyDelegate;
+        this.entity = (ThermalBankBlockEntity) blockEntity;
+
+        this.addSlot(new Slot(inv, 0, 104, 37));
+
+        addPlayerInventory(playerInventory);
+        addPlayerHotbar(playerInventory);
+
+        addProperties(propertyDelegate);
+    }
+
+    @Override
+    public ItemStack quickMove(PlayerEntity player, int invSlot) {
+        ItemStack newStack = ItemStack.EMPTY;
+        Slot slot = this.slots.get(invSlot);
+        if (slot != null && slot.hasStack()) {
+            ItemStack originalStack = slot.getStack();
+            newStack = originalStack.copy();
+            if (invSlot < this.inv.size()) {
+                if (!this.insertItem(originalStack, this.inv.size(), this.slots.size(), true)) {
+                    return ItemStack.EMPTY;
+                }
+            } else if (!this.insertItem(originalStack, 0, this.inv.size(), false)) {
+                return ItemStack.EMPTY;
+            }
+
+            if (originalStack.isEmpty()) {
+                slot.setStack(ItemStack.EMPTY);
+            } else {
+                slot.markDirty();
+            }
+        }
+
+        return newStack;
+    }
+
+    @Override
+    public boolean canUse(PlayerEntity player) {
+        return this.inv.canPlayerUse(player);
+    }
+
+    private void addPlayerInventory(PlayerInventory playerInventory) {
+        for (int i = 0; i < 3; ++i) {
+            for (int l = 0; l < 9; ++l) {
+                this.addSlot(new Slot(playerInventory, l + i * 9 + 9, 8 + l * 18, 84 + i * 18));
+            }
+        }
+    }
+
+    private void addPlayerHotbar(PlayerInventory playerInventory) {
+        for (int i = 0; i < 9; ++i) {
+            this.addSlot(new Slot(playerInventory, i, 8 + i * 18, 142));
+        }
+    }
+    public boolean isCrafting(){
+        return propertyDelegate.get(0) > 0;
+    }
+
+    public int getScaledProgress() {
+        int burnTime = this.propertyDelegate.get(0);
+        int fuelTime = this.propertyDelegate.get(1);
+        int progressArrowSize = 14;
+        return burnTime != 0 && fuelTime != 0 ? burnTime * progressArrowSize / fuelTime : 0;
+    }
+}